+++ src/spaces/controlPlaneV2/pages/McpPageV2.tsx
@@ -11,7 +11,7 @@ import {
 } from '@ui5/webcomponents-react';
 import { useParams, useSearchParams } from 'react-router-dom';
 import CopyKubeconfigButton from '../../../components/ControlPlanes/CopyKubeconfigButton.tsx';
-import styles from './McpPage.module.css';
+import styles from '../../mcp/pages/McpPage.module.css';
 // throws error sometimes if not imported
 import '@ui5/webcomponents-fiori/dist/illustrations/BeforeSearch';
 import { useTranslation } from 'react-i18next';
@@ -43,14 +43,14 @@ import { McpSecrets } from '../../../components/ControlPlane/McpSecrets.tsx';
 import { Providers } from '../../../components/ControlPlane/Providers.tsx';
 import { ProvidersConfig } from '../../../components/ControlPlane/ProvidersConfig.tsx';
 import Graph from '../../../components/Graphs/Graph.tsx';
-import { AuthProviderMcp } from '../auth/AuthContextMcp.tsx';
-import { ManagedControlPlaneAuthorization } from '../authorization/ManagedControlPlaneAuthorization.tsx';
-import { ComponentsDashboardV2 } from '../components/ComponentsDashboard/ComponentsDashboardV2.tsx';
-import { useCrossplaneQuery } from '../components/Kpi/useCrossplaneQuery.ts';
-import { useEsoQuery } from '../components/Kpi/useEsoQuery.ts';
-import { useFluxQuery } from '../components/Kpi/useFluxQuery.ts';
-import { useLandscaperQuery } from '../components/Kpi/useLandscaperQuery.ts';
-import { McpHeader } from '../components/McpHeader/McpHeader.tsx';
+import { AuthProviderMcp } from '../../mcp/auth/AuthContextMcp.tsx';
+import { ManagedControlPlaneAuthorization } from '../../mcp/authorization/ManagedControlPlaneAuthorization.tsx';
+import { ComponentsDashboardV2 } from '../../mcp/components/ComponentsDashboard/ComponentsDashboardV2.tsx';
+import { useCrossplaneQuery } from '../../mcp/components/Kpi/useCrossplaneQuery.ts';
+import { useEsoQuery } from '../../mcp/components/Kpi/useEsoQuery.ts';
+import { useFluxQuery } from '../../mcp/components/Kpi/useFluxQuery.ts';
+import { useLandscaperQuery } from '../../mcp/components/Kpi/useLandscaperQuery.ts';
+import { McpHeader } from '../../mcp/components/McpHeader/McpHeader.tsx';
 
 const MCP_PAGE_SECTIONS = ['overview', 'crossplane', 'flux', 'landscaper'] as const;
 export type McpPageSectionId = (typeof MCP_PAGE_SECTIONS)[number];
